requirements.txt
Django==2.1.4
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-elasticsearch-dsl==0.5.1
django-elasticsearch-dsl-drf==0.16.2
django-nine==0.1.13
djangorestframework==3.7.7
elasticsearch==6.0.0
elasticsearch-dsl==6.0.0
psycopg2==2.7.4
pytz==2018.3
six==1.11.0
urllib3==1.22
